Added Intro Scene Placeholder
Added a placeholder for our intro scene. Main menu button routes to placeholder intro scene which has a button that brings the user to the gameplay scene. This will be later changed to bring the user to the difficulty selection screen which will then bring the user to their specified gameplay screen.