Placeholder Buttons for Level Select and New Functionality for Difficulty Selection
I have created placeholders for the future blueprints on the Level Select screen. I have also imported the "Confirm" button from the drive into the project. Now the user can select a difficulty and artifact and press the confirm button to go to their desired level. The script I used is titled "LevelSelectButtonsScript". Sorry if this causes any confusion, I saw there were a couple other scripts that may have been intended to be used for this, but I was unsure who made them.